+++ src/main/java/org/spoutcraft/launcher/gui/LoginForm.java
@@ -20,7 +20,6 @@ import java.awt.Color;
 import java.awt.Component;
 import java.awt.Container;
 import java.awt.Dimension;
-import java.awt.FocusTraversalPolicy;
 import java.awt.Font;
 import java.awt.Graphics2D;
 import java.awt.Toolkit;
@@ -472,40 +471,6 @@ public class LoginForm extends JFrame implements ActionListener, DownloadListene
 	public void keyReleased(KeyEvent e) {
 	}
 
-	public static class SpoutFocusTraversalPolicy extends FocusTraversalPolicy {
-		Vector<Component> order;
-
-		public SpoutFocusTraversalPolicy(Vector<Component> order) {
-			this.order = new Vector<Component>(order.size());
-			this.order.addAll(order);
-		}
-
-		public Component getComponentAfter(Container focusCycleRoot, Component aComponent) {
-			int idx = (order.indexOf(aComponent) + 1) % order.size();
-			return order.get(idx);
-		}
-
-		public Component getComponentBefore(Container focusCycleRoot, Component aComponent) {
-			int idx = order.indexOf(aComponent) - 1;
-			if (idx < 0) {
-				idx = order.size() - 1;
-			}
-			return order.get(idx);
-		}
-
-		public Component getDefaultComponent(Container focusCycleRoot) {
-			return order.get(0);
-		}
-
-		public Component getLastComponent(Container focusCycleRoot) {
-			return order.lastElement();
-		}
-
-		public Component getFirstComponent(Container focusCycleRoot) {
-			return order.get(0);
-		}
-	}
-
 	private void readUsedUsernames() {
 		int i = 0;
 		try {
@@ -610,13 +575,15 @@ public class LoginForm extends JFrame implements ActionListener, DownloadListene
 	}
 
 	private void updatePasswordField() {
-		UserPasswordInformation info = usernames.get(this.usernameField.getSelectedItem().toString());
-		if (info.isHash) {
-			this.passwordField.setText("");
-			this.rememberCheckbox.setSelected(false);
-		} else {
-			this.passwordField.setText(info.password);
-			this.rememberCheckbox.setSelected(true);
+		if (this.usernameField.getSelectedItem() != null){
+			UserPasswordInformation info = usernames.get(this.usernameField.getSelectedItem().toString());
+			if (info.isHash) {
+				this.passwordField.setText("");
+				this.rememberCheckbox.setSelected(false);
+			} else {
+				this.passwordField.setText(info.password);
+				this.rememberCheckbox.setSelected(true);
+			}
 		}
 	}
